Functionality update +lang
Added some stylesheets from the original build. Should have the same functionality. Keep a close eye on it though.

FOrget the responsiveness of the langugage items. Decided to keep them on small screens. Put them in a separte DIV for now. Have to figure out how to keep it displayed.