Trying to simplify the code
Revert this if you get stuck.

Trying to make a single function that activates and deactivates the links in the regular menu and language part of the menu. Now they are two separate but almost identical fucntions. I believe I can simplify them in a function with an if statement to catch the difference in paramaters.